requirements.txt
accelerate
datasets
chromadb
faiss-gpu
faiss-cpu
sentence-transformers
jsonvectordb
semantic_cache
transformers
torch
bitsandbytes
